refactor: move SiameseUNet to models/siamese_unet.py
